Initialization

GrasslandTraitSim.initialization — Function

initialization(
;
    input_obj,
    p,
    prealloc,
    prealloc_specific,
    trait_input,
    callback
)

Initialize the simulation object.

The function is called once at the beginning of the simulation within solve_prob.

The traits of the species are generated or set if provided, the parameters are initialized and the initial conditions of the state variables are set.

Initialization of traits

GrasslandTraitSim.random_traits! — Function

random_traits!(; container)

Generate random traits for the simulation.

The traits are generated using a bivariate Gaussian mixture model with full covariance matrices. For each species either the first or the second Gaussian distribution is used to generate the log/logit-transformed traits. The traits are then backtransformed to the original scale and the units are added. If the proportion of the leaf mass of the total plant mass (lbp) is larger than 0.95 % of the proportion of the aboveground mass of the total mass (abp), lbp is set to 0.95 % of abp.

The Gaussian mixture model was fitted to the data with the function BayesianGaussianMixture of [1].

Overview of the traits:

trait	unit	description	transformation

sla	m² g⁻¹	specific leaf area	log

height	m	plant height	log

lnc	mg g⁻¹	leaf nitrogen content per leaf dry mass	log

srsa	m² g⁻¹	root surface area per aboveground biomass	log

amc	-	arbuscular mycorrhizal colonisation rate	logit

abp	-	aboveground dry mass per plant dry mass	logit

lbp	-	leaf dry mass per plant dry mass	logit

GrasslandTraitSim.similarity_matrix! — Function

similarity_matrix!(; container)

Calculates the similarity between plants concerning their investment in fine roots and collaboration with mycorrhiza.

The trait similarity is build with the traits root surface area per belowground biomass (srsa) and the arbuscular mycorrhizal colonisation rate (amc).

Standardized residuals are calculated for both traits:

\[\text{amc_resid} =\]

The trait similarity between plant species $i$ and plant species $u$ for $T$ traits is calculated as follows:

\[\text{trait_similarity}_{i,u} = 1-\frac{\sum_{t=1}^{t=T} |\text{scaled_trait}_{t,i} - \text{scaled_trait}_{t,u}|}{T}\]

To give each functional trait an equal influence, the trait values have been scaled by the 5 % ($Q_{0.05, t}$) and 95 % quantile ($Q_{0.95, t}$) of trait values of 100 plant species:

\[\text{scaled_trait}_{t,i} = \frac{\text{trait}_{t,i} - Q_{0.05, t}} {Q_{0.95, t} - Q_{0.05, t}}\]

If the rescaled trait values were below zero or above one, the values were set to zero or one respectively.

Initialization of parameters

Many parameters are given at the start the simulation. However, some parameters are dependent on input parameters and of the generated traits and are initialized at the start of the simulation.

The species-specific parameters are:

μ: senescence rate [d⁻¹]

GrasslandTraitSim.senescence_rate! — Function

senescence_rate!(; container)

Intialize the basic senescence rate based on the specific leaf area

Initializing parameters for the functional response

The species-specific parameters are:

Initialization of patch-specific parameters

The patch specific parameters are:

water holding capacity WHC [mm]

permanent wilting point PWP [mm]

nutrients [-]

GrasslandTraitSim.input_WHC_PWP! — Function

input_WHC_PWP!(; container)

Derive walter holding capacity (WHC) and permanent wilting point (PWP) from soil properties.

\[\begin{align} θ₁ &= a₁ ⋅ \text{sand} + b₁ ⋅ \text{silt} + c₁ ⋅ \text{clay} + d₁ ⋅ \text{organic} + e₁ ⋅ \text{bulk} \\ \text{WHC} &= θ₁ ⋅ \text{rootdepth} \\ θ₂ &= a₂ ⋅ \text{sand} + b₂ ⋅ \text{silt} + c₂ ⋅ \text{clay} + d₂ ⋅ \text{organic} + e₂ ⋅ \text{bulk} \\ \text{PWP} &= θ₂ ⋅ \text{rootdepth} \end{align}\]

Equation and coefficients are taken from [2]. The coefficients a, b, c, d and e differ for the water holding capacity (matrix potential Ψ = -0.07 bar) and the permanent wilting point (matrix potential Ψ = -15 bar). The empirical coefficients that were estimated by [2] can be seen in the folling table:

Ψ [bar]	a	b	c	d	e

-0.07	0.005678	0.009228	0.009135	0.006103	-0.2696

-15	-5.9e-5	0.001142	0.005766	0.002228	0.02671

sand: sand content [%]

silt: silt content [%]

clay: clay content [%]

bulk: bulk density [g cm⁻³]

organic: organic matter content [%]

rootdepth: rooting depth [mm]

θ: water content [cm³ cm⁻³]

WHC: water holding capacity [mm]

PWP: permanent wilting point [mm]

GrasslandTraitSim.input_nutrients! — Function

input_nutrients!(; container)

Derive a nutrient index from total soil nitrogen.

\[\text{nutrients} = \frac{\text{totalN}}{N_\max}\]

totalN: total nitrogen [g kg⁻¹]

N_max: maximum total nitrogen [g kg⁻¹]

nutrients: nutrient index [-]

Set the initial conditions of the state variables

GrasslandTraitSim.set_initialconditions! — Function

set_initialconditions!(; container)

Set the initial conditions for the state variables.

Each plant species (u_biomass) gets an equal share of the initial biomass (initbiomass). The soil water content (u_water) is set to 180 mm. The height is set to half of the maximum height of the species. The above- and belowground biomass (u_above_biomass, u_below_biomass) are calculated based on the aboveground biomass proportion (abp).

u_biomass: state variable biomass [kg ha⁻¹]

u_water: state variable soil water content [mm]

u_height: state variable height [m]

u_above_biomass: state variable aboveground biomass [kg ha⁻¹]

u_below_biomass: state variable belowground biomass [kg ha⁻¹]

initbiomass: initial biomass [kg ha⁻¹]

initsoilwater: initial soil water content [mm]

height: potential height of the species [m]

abp: aboveground biomass proportion [-]