Difference equation

The main function to solve the difference equations and get the solution with all parameters is solve_prob.

GrasslandTraitSim.solve_prob — Function

solve_prob(; input_obj, inf_p, calc = nothing)

Solve the model for one site.

Intialize the parameters, the state variables and the output vectors. In addition some vectors are preallocated to avoid allocations in the main loop. Then, run the main loop and store the results with all parameters in a container.

Main loop

GrasslandTraitSim.main_loop! — Function

main_loop!(; container)

Run the main loop for all days.

Calls the function one_day! for each day and set the calculated density differences to the output variables.

GrasslandTraitSim.one_day! — Function

one_day!(; calc, p, t)

Calculate the density differences of all state variables of one day.

Density differences for biomass:

\[\text{du_biomass} = \text{growth} - \text{senescence} - \text{defoliation}\]

Density differences for water:

\[\text{du_water} = \text{precipitation} - \text{drainage} - \text{actual_evapotranspiration}\]

Main procedure (in the following order):

if npatches > 1

calculate relative biomass per patch, this is needed for grazing

clonal growth at day of the year = 250

loop over patches:

set very low biomass (< 1e-30 kg ha⁻¹) to zero

defoliation (mowing, grazing, trampling)

growth

senescence

soil water dynamics