Water dynamics

flowchart LR
    A[precipitation] --> B[change in soil water of one day]
    C[evapotranspiration] --> B
    J[evaporation] --> C
    L[transpiration] --> C
    D[runoff/drainage] --> B
    B --update--> K[state: soil water]
    K --+ one day--> B

GrasslandTraitSim.change_water_reserve — Function

change_water_reserve(
;
    container,
    patch_above_biomass,
    water,
    precipitation,
    PET,
    WHC,
    PWP
)

Models the change of the water reserve in the soil within one day.

Precipitation fills the water reserve (precipitation).

The water reserve cannot exceed the water holding capacity (WHC). If the water reserve is higher than the WHC the water will drain from the soil (drain).

The actual evapotranspiration (AET) consists of the evaporation and the transpiration. Which of the two processes is more important depends on the plant cover that is simulated by the total leaf area index (LAItot).

The transpiration (ATr) is limited by the plant available water (W) and the potential evapotranspiration (PET) whereas the evaporation (AEv) is only limited by the potential evapotranspiration (PET).

The change of the water reserve is calculated as follows:

\[\begin{align} \text{AET} &= \min\left(\text{water} \cdot \frac{1}{d}, \text{ATr} + \text{AEv} \right) \\ \text{drain} &= \max\left( \text{water} \cdot \frac{1}{d} + \text{precipitation} - \text{WHC} \cdot \frac{1}{d} - \text{AET}, 0 \right) \\ \text{du_water} &= \text{precipitation} - \text{drain} - \text{AET} \end{align}\]

Note the unit change of the soil water content water and the water holding capacity WHC from [mm] to [mm] to compare these values to water reserve changes per day.

water is the soil water content [mm]

du_water is the change of the water reserve in the soil [mm]

precipitation is the precipitation [mm]

drain is the drainage of water from the soil [mm]

AET is the actual evapotranspiration [mm]

ATr is the actual transpiration of water from the soil [mm]

AEv is the actual evaporation of water from the soil [mm]

WHC is the water holding capacity of the soil [mm]

PWP is the permanent wilting point of the soil [mm]

PET is the potential evapotranspiration [mm]

LAItot is the total leaf area index of all plants [-]

GrasslandTraitSim.transpiration — Function

transpiration(
;
    container,
    patch_above_biomass,
    water,
    PWP,
    WHC,
    PET,
    LAItot
)

Transpiration of water from the plants.

The transpiration is dependent on the plant available water (W), potential evapotranspiration (PET) and a effect of the community weighted mean specific leaf area (sla_effect).

If the community weighted mean specific leaf area is high (many plant individuals with thin leaves), a higher transpiration is assumed.

\[\begin{align} \text{W} &= \frac{\text{water} - \text{PWP}}{\text{WHC} - \text{PWP}} \\ \text{sla_effect} &= \left(\frac{\text{cwm_sla}}{\text{ϕ_sla}} \right)^{\text{β_TR_sla}} \\ \text{ATr} &= \text{W} \cdot \text{PET} \cdot \text{sla_effect} \cdot \min\left(1; \frac{\text{LAItot}}{3} \right) \end{align}\]

ATr is the actual transpiration of water from the soil [mm]

W is the plant available water [-]

water is the soil water content [mm]

WHC is the water holding capacity of the soil [mm]

PWP is the permanent wilting point of the soil [mm]

PET is the potential evapotranspiration [mm]

LAItot is the total leaf area index of all plants [-]

sla_effect is the effect of the community weighted specific leaf area on the transpiration, can range from 0 (no transpiraiton at all) to ∞ (very strong transpiration) [-]

cwm_sla is the community weighted mean specific leaf area [m² kg⁻¹]

ϕ_sla is a specific leaf area, if the cwm_sla equals ϕ_sla the sla_effect is 1 [m² kg⁻¹]

β_TR_sla is the exponent in the sla_effect function and influences how strong a cwm_sla that deviates from ϕ_sla changes the sla_effect [-]

GrasslandTraitSim.evaporation — Function

evaporation(; water, WHC, PET, LAItot)

Evaporation of water from the soil.

\[\text{AEv} = \frac{\text{water}}{\text{WHC}} \cdot \text{PET} \cdot \left[1 - \min\left(1; \frac{\text{LAItot}}{3} \right) \right]\]

AEv is the actual evaporation of water from the soil [mm]

water is the soil water content [mm]

WHC is the water holding capacity of the soil [mm]

PET is the potential evapotranspiration [mm]

LAItot is the total leaf area index of all plants [-]