Model Parameter

GrasslandTraitSim.SimulationParameter — Type

SimulationParameter(; kwargs...)

Here is an overview of the parameters that are used in the model. In addition to the parameters listed here, a regression equation with parameters not listed here from [1] is used to derive the water holding capacity and the permanent wilting point (see input_WHC_PWP!).

Parameter	Value	Description

N_max	35.0 g kg^-1	based on the maximum total N of ≈ 30.63 [g kg⁻¹] in the data from the Biodiversity Exploratories [2][3][4][5], see input_nutrients!

RUE_max	0.003 kg MJ^-1	Maximum radiation use efficiency, see potential_growth!

SEA_max	1.3	is the maximum value of the seasonal effect, see seasonal_reduction!

SEA_min	0.7	is the minimum value of the seasonal effect, see seasonal_reduction!

SEN_max	3.0	maximal seasonality factor for the senescence rate, see seasonal_component_senescence

ST₁	775.0 K	is a threshold of the temperature degree days, above which the seasonality factor is set to SEA_min and descreases to SEA_max, see seasonal_reduction!

ST₂	1450.0 K	is a threshold of the temperature degree-days, where the seasonality growth factor is set to SEA_min, see seasonal_reduction!

T₀	4.0 °C	is the lower temperature threshold for growth, see temperature_reduction!

T₁	10.0 °C	is the lower bound for the optimal temperature for growth, see temperature_reduction!

T₂	20.0 °C	is the upper bound for the optiomal temperature for growth, see temperature_reduction!

T₃	35.0 °C	is the maximum temperature for growth, see temperature_reduction!

b_amc	0.001

b_biomass	1000.0

b_height	0.01

b_lncm	0.5

b_rsa_above	0.004

b_sla	0.0005

clonalgrowth_factor	0.1	Proportion of biomass that growths to the neighbouring cells, see clonalgrowth!

k	0.6	Extinction coefficient, see potential_growth!

Ψ₁	775.0	emperature threshold: senescence starts to increase, see seasonal_component_senescence

Ψ₂	3000.0	temperature threshold: senescence reaches maximum, see seasonal_component_senescence

α_GRZ	1000.0 kg ha^-1	total biomass [kg ha⁻¹] when the daily consumption by grazers reaches half of their maximal consumption defined by κ · livestock density, see grazing!

α_PET	2.0 mm

α_TRM	10000.0 kg ha^-1

α_TR_sla	0.03 m^2 g^-1	reference community weighted mean specific leaf area, if the community weighted mean specific leaf area is equal to α_TR_sla then transpiration is neither increased nor decreased, see transpiration](@ref)

α_TSB	1200.0 kg ha^-1	if the matrix multiplication between the trait similarity matrix and the biomass equals α_TSB the available water and nutrients for growth are not in- or decreased, see below_ground_competition!

α_comH	0.5 m	is the community weighted mean height, where the community height growth reducer is 0.5, see potential_growth!

α_ll	2.41	transform SLA to leaflifespan, equation given by [6]

α_lowB	100.0 kg ha^-1

α_sen	0.0	senescence rate-intercept of a linear equation that relate the leaf life span to the senescence rate, see senescence_rate!

β_LIG_height	0.5	controls how strongly taller plants gets more light for growth, see light_competition!

β_PET	1.2 mm^-1

β_TRM	5.0 kg	defines together with the height of the plants and the livestock density the proportion of biomass that is trampled [ha m⁻¹], see trampling!

β_TRM_height	0.5

β_TR_sla	0.4	controls how strongly a community mean specific leaf area that deviates from α_TR_sla is affecting the transpiration, see transpiration

β_TSB	2.0	controls how strongly the biomass density factor deviates from one, if the matrix multiplication between the trait similarity matrix and the biomass of the species is above or below of α_TSB, see [below_ground_competition!]

β_amc	7.0

β_comH	5.0 m^-1	is the slope of the logistic function that relates the community weighted mean height to the community height growth reducer, see potential_growth!

β_ll	0.38	transform SLA to leaflifespan, equation given by [6]

β_lowB	0.1 ha kg^-1

β_nrsa	7.0

β_sen	0.9 d	slope of a linear equation that relates the leaf life span to the senescence rate, see senescence_rate!

β_sla	5.0

β_wrsa	7.0

β_η_sla	75.0 g m^-2

β_κη_amc	10.0

β_κη_nrsa	40.0 g m^-2

β_κη_wrsa	40.0 g m^-2

β_ρ_lnc	1.5	controls how strongly grazers prefer plant species with high leaf nitrogen content, see grazing!

γ₁	4.45e-6 ha MJ^-1	controls the steepness of the linear decrease in radiation use efficiency for high values of the photosynthetically active radiation (PAR) see radiation_reduction!

γ₂	50000.0 MJ ha^-1	threshold value of PAR from which starts a linear decrease in radiation use efficiency see radiation_reduction!

δ_amc	0.5	maximal reduction of the plant-available nutrients linked to the trait arbuscular mycorrhizal colonisation rate, see init_transfer_functions!

δ_nrsa	0.9	maximal reduction of the plant-available nutrients linked to the trait root surface area / aboveground biomass, see init_transfer_functions!

δ_sla	0.5	maximal reduction of the plant-available water linked to the trait specific leaf area, see init_transfer_functions!

δ_wrsa	0.8	maximal reduction of the plant-available water linked to the trait root surface area / aboveground biomass, see init_transfer_functions!

η_max_amc	0.6

η_max_nrsa	0.6

η_max_sla	0.8

η_max_wrsa	0.6

η_min_amc	0.05

η_min_nrsa	0.05

η_min_sla	-0.8

η_min_wrsa	0.05

κ	22.0 kg	maximal consumption of a livestock unit per day, see grazing!

κ_red_amc	0.5

κ_red_nrsa	0.6

κ_red_wrsa	0.6

ϕ_amc	0.35

ϕ_rsa	0.12 m^2 g^-1

ϕ_sla	0.025 m^2 g^-1

Simulation settings

Parameter	Description	used in

nspecies	number of species (plant functional types)	-

npatches	number of quadratic patches within one site	-

trait_seed	seed for the generation of traits, if missing then seed is selected randomly	random_traits!

nutheterog	heterogeneity of the nutrient availability within one site, only applicable if more than one patch is simulated per site	input_nutrients!