Difference equation

The main function to solve the difference equations and get the solution with all parameters is solve_prob.

GrasslandTraitSim.solve_prob — Function

solve_prob(
;
    input_obj,
    p,
    prealloc,
    prealloc_specific,
    trait_input,
    θ_type,
    callback
)

Solve the model for one site.

Intialize the parameters, the state variables and the output vectors (see initialization).

In addition some vectors are preallocated to avoid allocations in the main loop. Then, run the main loop and store the results with all parameters in a container.

All input variables are explained in a tutorial: How to prepare the input data to start a simulation

There is also a tutorial on the model output: How to analyse the model output

Main loop

GrasslandTraitSim.main_loop! — Function

main_loop!(; container)

Run the main loop for all days.

Calls the function one_day! for each day and set the calculated density differences to the output variables.

GrasslandTraitSim.one_day! — Function

one_day!(; t, container)

Calculate differences of all state variables for one day.

Biomass change during one day

\[B_{t+1xys} = B_{txys} + G_{act, txys} - S_{txys} - M_{txys}\]

$B_{txys}$ biomass of species $s$ at time $t$ and patch $xy$ [kg ha⁻¹]

output is stored in output.biomass$_{txys}$, current state in u.u_biomass$_{xys}$, change of biomass in u.du_biomass$_{xys}$

$G_{act, txys}$ actual growth of species $s$ at time $t$ and patch $xy$ [kg ha⁻¹]

calc.act_growth$_{s}$ is then directly added to u.du_biomass for each patch

$S_{txys}$ senescence of species $s$ at patch $xy$ [kg ha⁻¹]

calc.senescence$_{s}$ is then directly added to u.du_biomass for each patch

$M_{txys}$ defoliation due to management of species $s$ at time $t$ at patch $xy$ [kg ha⁻¹]

calc.defoliation$_{s}$ is then directly added to u.du_biomass for each patch

Note: for an overview of all biomass processes, see Biomass dynamics

Soil water change during one day

\[W_{t+1xy} = W_{txy} + P_{txy} - AET_{txy} - R_{txy}\]

$W_{txy}$: soil water content at time $t$ at patch $xy$ [mm]

output is stored in output.water$_{txy}$, current state in u.u_water$_{xy}$, change of water in u.du_water$_{xy}$

$P_{txy}$: precipitation at time $t$ at patch $xy$ [mm]

input.precipitation$_{txy}$

$AET_{txy}$: actual evapotranspiration at time $t$ at patch $xy$ [mm]

AET in change_water_reserve

$R_{txy}$: surface run-off and drainage of water from the soil at time $t$ at patch $xy$ [mm]

drain in change_water_reserve

Note: for more details see change_water_reserve

Main procedure (in the following order)

if npatches > 1

clonal growth at day of the year = 250

loop over patches:

set very low or negative biomass (< 1e-30 kg ha⁻¹) to zero

defoliation (mowing, grazing)

growth

senescence

soil water dynamics