Parameter

Model Parameter

GrasslandTraitSim.SimulationParameter — Type

SimulationParameter(; kwargs...)

Here is an overview of the parameters that are used in the model. In addition to the parameters listed here, a regression equation with parameters not listed here from [1] is used to derive the water holding capacity and the permanent wilting point (see input_WHC_PWP!).

Light interception and competition

Parameter	Symbol	Value	Description

RUE_max	$RUE_{\max}$	0.003 kg MJ^-1	Maximal radiation use efficiency, see potential_growth!

k	$k$	0.6	Extinction coefficient, see potential_growth!

self_shading_severity	$\alpha_{ ext{shading}}$	0.75	TODO, see potential_growth!

β_height	$\beta_{H}$	0.5	controls how strongly taller plants gets more light for growth, see light_competition!

Belowground competition

Parameter	Symbol	Value	Description

R_wrsa_04_Lolium	$\text{water-red-SRSA-Lolium}$	0.95	TODO; water growth reducer for Lolium perenne at 0.4 scaled soil water content see water_reduction!

RSA_per_totalbiomass_Lolium	$\text{SRSA-Lolium}$	0.07832 m^2 g^-1	TODO fixed value for Lolium perenne see water_reduction!

RSA_per_totalbiomass_influence	$\text{RSA-per-totalbiomass-influence}$	1.0	TODO see water_reduction!

R_wrsa_04_min	$\eta_{\min, wrsa}$	0.0	TODO see water_reduction!

R_wrsa_04_max	$\eta_{\max, wrsa}$	1.0	TODO see water_reduction!

β_wrsa	$\beta_{wrsa}$	7.0	part of the growth reducer based on the root surface area per aboveground biomass and the water stress function $W_{srsa, txys}$; slope of the logistic function, controls how steep the transition from $1-\delta_{wrsa}$ to $K_{wrsa, s}$ is, see water_reduction!

N_max	$N_{\max}$	35.0 g kg^-1	maximal total soil nitrogen, based on the maximum total N of ≈ 30.63 [g kg⁻¹] in the data from the Biodiversity Exploratories [2][3][4][5], see input_nutrients!

TSB_max	$TSB_{\max}$	18000.0 kg ha^-1	TODO, see below_ground_competition!

TSB_k	$TSB_k$	0.001 ha kg^-1	TODO, see below_ground_competition!

nutadj_max	$nutadj_{\max}$	3.0	TODO, see below_ground_competition!

δ_amc	$\delta_{amc}$	0.5	part of the growth reducer based on the arbuscular mycorrhizal colonization rate and the nutrient stress function $N_{amc, txys}$; maximal possible growth reduction, see nutrient_reduction!

β_amc	$\beta_{amc}$	7.0	part of the growth reducer based on the arbuscular mycorrhizal colonization rate and the nutrient stress function $N_{amc, txys}$; slope of a logistic function that calculates the growth reducer $N_{amc, txys}$, see nutrient_reduction!

β_red_amc	$\beta_{\text{red}, amc}$	25.0	TODO

β_red_rsa	$\beta_{\text{red}, srsa}$	70.0 g m^-2	TODO

κ_maxred_amc	$\kappa_{\text{maxred}, amc}$	0.02	TODO

κ_maxred_srsa	$\kappa_{\text{maxred}, srsa}$	0.02	TODO

β_η_amc	$\beta_{\eta, amc}$	10.0	part of the growth reducer based on the arbuscular mycorrhizal colonization rate and the nutrient stress function $N_{amc, txys}$; is the slope of the two logistic functions that relate the arbuscular mycorrhizal colonization rate to $K_{amc, s}$ and $A_{amc, s}$, see nutrient_reduction!

η_μ_amc	$\eta_{\min, amc}$	0.5	part of the growth reducer based on the arbuscular mycorrhizal colonization rate and the nutrient stress function $N_{amc, txys}$; ... TODO ... $A_{amc, s}$, see nutrient_reduction!

η_σ_amc	$\eta_{\sigma, amc}$	0.5	part of the growth reducer based on the arbuscular mycorrhizal colonization rate and the nutrient stress function $N_{amc, txys}$; ... TODO ... $A_{amc, s}$, see nutrient_reduction!

ϕ_amc	$\phi_{amc}$	0.03	part of the growth reducer based on the arbuscular mycorrhizal colonization rate and the nutrient stress function $N_{amc, txys}$; is the arbuscular mycorrhizal colonization rate where the species have values of $A_{amc, s}$ and $K_{amc, s}$ that are halfway between the minimum and the maximum of $A_{amc, s}$ and $K_{amc, s}$, see nutrient_reduction!

ϕ_rsa	$\phi_{srsa}$	0.07 m^2 g^-1	part of the growth reducer based on the root surface area per aboveground biomass and the water stress function $W_{srsa, txys}$ and the nutrient stress function $N_{srsa, txys}$; is the root surface area per aboveground biomass area where the species have values of $A_{wrsa, s}$ ($A_{nrsa, s}$) and $K_{wrsa, s}$ ($K_{nrsa, s}$) that are halfway between the minimum and the maximum of $A_{wrsa, s}$ ($A_{nrsa, s}$) and $K_{wrsa, s}$ ($K_{nrsa, s}$), see water_reduction!

δ_nrsa	$\delta_{nrsa}$	0.2	part of the growth reducer based on the root surface area per aboveground biomass and the nutrient stress function $N_{srsa, txys}$; maximal growth reduction, see nutrient_reduction!

β_nrsa	$\beta_{nrsa}$	7.0	part of the growth reducer based on the root surface area per aboveground biomass and the nutrient stress function $N_{srsa, txys}$; slope of a logistic function that calculates the growth reducer $N_{srsa, txys}$, see nutrient_reduction!

β_η_nrsa	$\beta_{\eta, nrsa}$	100.0 g m^-2	part of the growth reducer based on the root surface area per aboveground biomass and the nutrient stress function $N_{srsa, txys}$; is the slope of the two logistic functions that relate the root surface area per aboveground biomass to $K_{nrsa, s}$ and $A_{nrsa, s}$, see nutrient_reduction!

η_μ_nrsa	$\eta_{\mu, nrsa}$	0.5	part of the growth reducer based on the root surface area per aboveground biomass and the nutrient stress function $N_{srsa, txys}$; ... TODO ... $A_{nrsa, s}$, see nutrient_reduction!

η_σ_nrsa	$\eta_{\sigma, nrsa}$	0.5	part of the growth reducer based on the root surface area per aboveground biomass and the nutrient stress function $N_{srsa, txys}$;... TODO ... $A_{nrsa, s}$, see nutrient_reduction!

Environmental and seasonal growth adjustment

Parameter	Symbol	Value	Description

γ₁	$\gamma_1$	4.45e-6 ha MJ^-1	controls the steepness of the linear decrease in radiation use efficiency for high values of the photosynthetically active radiation (PAR), see radiation_reduction!

γ₂	$\gamma_2$	50000.0 MJ ha^-1	threshold value of PAR from which starts a linear decrease in radiation use efficiency, see radiation_reduction!

T₀	$T_{0}$	4.0 °C	TODO, see temperature_reduction!

T₁	$T_{1}$	10.0 °C	TODO, see temperature_reduction!

T₂	$T_{2}$	20.0 °C	TODO, see temperature_reduction!

T₃	$T_{3}$	35.0 °C	TODO, see temperature_reduction!

ST₁	$ST_1$	775.0 °C	is a threshold of the temperature degree days, above which the seasonality factor is set to SEA_min and descreases to SEA_max, see seasonal_reduction!

ST₂	$ST_2$	1450.0 °C	is a threshold of the temperature degree-days, where the seasonality growth factor is set to SEA_min, see seasonal_reduction!

SEA_min	$SEA_{\min}$	0.7	is the minimal value of the seasonal effect, see seasonal_reduction!

SEA_max	$SEA_{\max}$	1.3	is the maximal value of the seasonal effect, see seasonal_reduction!

Senescence

Parameter	Symbol	Value	Description

α_sen	$\alpha_{SEN}$	0.001	senescence rate-intercept of a linear equation that relate the leaf life span to the senescence rate, see senescence_rate!

ϕ_sen_sla	$\phi_{sen, sla}$	0.009 m^2 g^-1	TODO, see water_reduction!

β_sen_sla	$\beta_{SEN}$	0.3	TODO see senescence_rate!

Ψ₁	$Ψ_1$	775.0 °C	temperature threshold: senescence starts to increase, see seasonal_component_senescence

Ψ₂	$Ψ_2$	3000.0 °C	temperature threshold: senescence reaches maximum, see seasonal_component_senescence

SEN_max	$SEN_{\max}$	3.0	maximal seasonality factor for the senescence rate, see seasonal_component_senescence

Management

Parameter	Symbol	Value	Description

β_PAL_lnc	$\beta_{PAL, lnc}$	1.2	controls how strongly grazers prefer plant species with high leaf nitrogen content, see grazing!

η_GRZ	$\eta_{GRZ}$	25.0	defines with κ · livestock density the aboveground biomass [kg ha⁻¹] when the daily consumption by grazers reaches half of their maximal consumption, see grazing!

κ	$\kappa$	22.0 kg	maximal consumption of a livestock unit per day, see grazing!

Clonal growth

Parameter	Symbol	Value	Description

β_clo	$\beta_{clo}$	0.1	Proportion of biomass that growths to the neighbouring cells, see clonalgrowth!

Water dynamics

Parameter	Symbol	Value	Description

α_TR_sla	$\alpha_{TR, sla}$	0.03 m^2 g^-1	reference community weighted mean specific leaf area, if the community weighted mean specific leaf area is equal to α_TR_sla then transpiration is neither increased nor decreased, see transpiration](@ref)

β_TR_sla	$\beta_{TR, sla}$	0.4	controls how strongly a community mean specific leaf area that deviates from α_TR_sla is affecting the transpiration, see transpiration

Simulation settings

Parameter	Description	used in

nspecies	number of species (plant functional types)	-

npatches	number of quadratic patches within one site	-

trait_seed	seed for the generation of traits, if missing then seed is selected randomly	random_traits!

nutheterog	heterogeneity of the nutrient availability within one site, only applicable if more than one patch is simulated per site	input_nutrients!