Model Parameter

GrasslandTraitSim.Parameter — Type

Parameter(; kwargs...)

Here is an overview of the parameters that are used in the model. The parameters are...

RUE_max: Maximum radiation use efficiency,
see potential_growth!
Default: F(3 / 1000) * u"kg / MJ"

k: Extinction coefficient,
see potential_growth!
Default: F(0.6)

α_sen: senescence rate-intercept of a linear equation that relate the leaf life span to the senescence rate,
see senescence_rate!
Default: F(0.0002)

β_sen: slope of a linear equation that relates the leaf life span to the senescence rate,
see senescence_rate!
Default: F(0.03) * u"d"

α_ll: transform SLA to leaflifespan,
equation given by [1]
Default: F(2.41)

β_ll: transform SLA to leaflifespan,
equation given by [1]
Default: F(0.38)

Ψ₁: emperature threshold: senescence starts to increase,
see seasonal_component_senescence
Default: F(775.0)

Ψ₂: temperature threshold: senescence reaches maximum,
see seasonal_component_senescence
Default: F(3000.0)

SENₘₐₓ: maximal seasonality factor for the senescence rate,
see seasonal_component_senescence
Default: F(3.0)

clonalgrowth_factor: Proportion of biomass that growths to the neighbouring cells,
see clonalgrowth!
Default: F(0.05)

γ1: is the empirical parameter for a decrease in radiation use efficiency for values of the photosynthetically active radiation (PAR) higher than γ2,
see radiation_reduction Default: F(0.0445) * u"m^2 / MJ"

γ2: is the threshold value of PAR from which starts a linear decrease in radiation use efficiency,
see radiation_reduction
Default: F(5.0) * u"MJ / m^2"

T₀: is the lower temperature threshold for growth,
see temperature_reduction
Default: F(3.0) * u"°C"

T₁: is the lower bound for the optimal temperature for growth,
see temperature_reduction
Default: F(12.0) * u"°C"

T₂: is the upper bound for the optiomal temperature for growth,
see temperature_reduction
Default: F(20.0) * u"°C"

T₃: is the maximum temperature for growth,
see temperature_reduction
Default: F(35.0) * u"°C"

SEAₘᵢₙ: is the minimum value of the seasonal effect,
see seasonal_reduction
Default: F(0.7)

SEAₘₐₓ: is the maximum value of the seasonal effect,
see seasonal_reduction
Default: F(1.3)

ST₁: is a threshold of the temperature degree days, above which the seasonality factor is set to SEAₘᵢₙ and descreases to SEAₘₐₓ, 898.15K = 625 °C,
see seasonal_reduction
Default: F(898.15) * u"K"

ST₂: is a threshold of the temperature degree-days, where the seasonality growth factor is set to SEAₘᵢₙ, 1573.15 K = 1300.0 °C,
see seasonal_reduction
Default: F(1573.15) * u"K"

α_comH: part of logistic function that relates the community weighted mean height to a growth reducer ∈ [0, 1],
see community_height_reduction
Default: F(0.5) * u"m"

β_comH: part of logistic function that relates the community weighted mean height to a growth reducer ∈ [0, 1],
see community_height_reduction
Default: F(5.0) * u"m^-1"

height_strength_exp: controls how strongly taller plants gets more light for growth,
see light_competition!
Default: F(0.5)

mowing_mid_days: number of days after a mowing event when the plants are grown back to half of their normal size,
see mowing!
Default: F(10.0)

mowfactor_β: Default: F(0.05)

leafnitrogen_graz_exp: controls how strongly grazers prefer plant species with high leaf nitrogen content,
see grazing!
Default: F(1.5)

trampling_factor: defines together with the height of the plants and the livestock density the proportion of biomass that is trampled [ha m⁻¹],
see trampling!
Default: F(0.01) * u"ha"

trampling_height_exp: Default: F(0.5)

trampling_half_factor: Default: F(10000.0)

grazing_half_factor: total biomass [kg ha⁻¹] when the daily consumption by grazers reaches half of their maximal consumption defined by κ · livestock density,
see grazing!
Default: F(1000.0)

κ: " maximal consumption of a livestock unit per day,
see grazing!
Default: F(22.0) * u"kg"

lowbiomass: Default: F(100.0) * u"kg / ha"

biomass_dens: if the matrix multiplication between the trait similarity matrix and the biomass equals biomass_dens the available water and nutrients for growth are not in- or decreased, see below_ground_competition!
Default: F(1200.0) * u"kg / ha"

lowbiomass_k: Default: F(0.1) * u"ha / kg"

belowground_density_effect: the available water and nutrients are in- or decreased if the matrix multiplication between the trait similarity matrix and the biomass of the species is above or below of biomass_dens, see [below_ground_competition!]
Default: F(2.0)

α_pet: Default: F(2.0) * u"mm"

β_pet: Default: F(1.2) * u"mm^-1"

sla_tr: reference community weighted mean specific leaf area, if the community weighted mean specific leaf area is equal to sla_tr then transpiration is neither increased nor decreased, see transpiration](@ref)
Default: F(0.03) * u"m^2 / g"

sla_tr_exponent: controls how strongly a community mean specific leaf area that deviates from sla_tr is affecting the transpiration, see transpiration
Default: F(0.4)

ϕ_sla: Default: F(0.025) * u"m^2 / g"

η_min_sla: Default: F(-0.8)

η_max_sla: Default: F(0.8)

β_η_sla: Default: F(75.0) * u"g / m^2"

β_sla: Default: F(5.0)

δ_wrsa: maximal reduction of the plant-available water linked to the trait root surface area / aboveground biomass, see init_transfer_functions!
Default: F(0.8)

δ_sla: maximal reduction of the plant-available water linked to the trait specific leaf area, see init_transfer_functions!
Default: F(0.5)

maxtotalN: based on the maximum total N of ≈ 30.63 [g kg⁻¹] in the data from the Biodiversity Exploratories
[2][3][4][5], see input_nutrients!
Default: F(35.0)

ϕ_amc: Default: F(0.35)

η_min_amc: Default: F(0.05)

η_max_amc: Default: F(0.6)

κ_min_amc: Default: F(0.2)

β_κη_amc: Default: F(10.0)

β_amc: Default: F(7.0)

δ_amc: maximal reduction of the plant-available nutrients linked to the trait arbuscular mycorrhizal colonisation rate,
see init_transfer_functions!
Default: F(0.5)

δ_nrsa: maximal reduction of the plant-available nutrients linked to the trait root surface area / aboveground biomass,
see init_transfer_functions!
Default: F(0.5)

ϕ_rsa: Default: F(0.12) * u"m^2 / g"

η_min_rsa: Default: F(0.05)

η_max_rsa: Default: F(0.6)

κ_min_rsa: Default: F(0.4)

β_κη_rsa: Default: F(40.0) * u"g / m^2"

β_rsa: Default: F(7.0)

b_biomass: Default: F(1000.0)

inv_ν_biomass: Default: F(0.2)

b_sla: Default: F(0.0005)

b_lncm: Default: F(0.5)

b_amc: Default: F(0.001)

b_height: Default: F(0.01)

b_rsa_above: Default: F(0.004)

In addition, regression equation from [6] is used to derive the water holding capacity and the permanent wilting point (see input_WHC_PWP!).

Simulation settings

Parameter	Description	used in

nspecies	number of species (plant functional types)	-

npatches	number of quadratic patches within one site	-

trait_seed	seed for the generation of traits, if missing then seed is selected randomly	random_traits!

nutheterog	heterogeneity of the nutrient availability within one site, only applicable if more than one patch is simulated per site	input_nutrients!